Added Syllabus to Files array
Reworked the course files array; rather than having everything as
sample, just changed it to "good","bad","syllabus", so that I could just
add the Syllabus PDF file to the array. Trying to figure out how to add
the course name to the files as well, which is proving a challenge.
Converting PDFs to HTML to add to the description.