Adjust Syllabi
Received notice from Professor Kelly, able to only use degree plan classes. Removed extraneous classes, will build array based on all ET classes in the degree plan, if no syllabus is given will mark so on the title of the page and in the description. Yay for going from 50+ courses to 20!